Number the docs for each step
During the previous tutorial, folks found it hard to navigate
between the steps because the file names did not indicate the
order the steps were in.

Updated mock tests to make use of pytest-mock, which is
the recommended way of mixing mocking and pytest.

Made step 5 and 7 consistent with the rest of the steps.

Added links to solutions, looks like step 7 doesn't have
any solutions for folks to refer to.

Also added a couple pictures to the steps to make them
a bit more fun.