Make docs clearer and requirements less restrictive
Very restrictive requirements can make it difficult if you
already have some of these packages installed on your machine,
for example if you already have pytest installed but it doesn't
happen to be the exact version we're looking for.

In Step 1 made it clear that you need to be in the directory
of the git repo you just checked out before creating a branch.