Add refatoring slide example
This is the example that will be in our slides for refatoring
for unit testability.

This example shows a method `setup_cats` which is doing several
things. The refactored version still does 2 things, however it
is gluing together separate single responsibilty functions
(arguably it would be a cleaner interface to have two functions
called by the user: one a factory for cats, one to feed cats!)